+++ src/sphere.cpp
@@ -9,18 +9,18 @@ std::vector<Intersection> Sphere::Intersect(Ray r) {
     Tuple sphere_to_ray = r2.origin - Point(0,0,0);
 
     float a = r2.direction.Dot(r2.direction);
-    float b = 2 * r2.direction.Dot(sphere_to_ray);
-    float c = sphere_to_ray.Dot(sphere_to_ray) - 1;
+    float b = 2.f * r2.direction.Dot(sphere_to_ray);
+    float c = sphere_to_ray.Dot(sphere_to_ray) - 1.f;
 
-    float discriminant = pow(b, 2) - (4 * a * c);
+    float discriminant = powf(b, 2) - (4.f * a * c);
 
     // IF there's no intersections
-    if(discriminant < 0) {
+    if(discriminant < 0.f) {
         return std::vector<Intersection> {};
     }
 
-    float t1 = (-b - sqrt(discriminant)) / (2 * a);
-    float t2 = (-b + sqrt(discriminant)) / (2 * a);
+    float t1 = (-b - sqrtf(discriminant)) / (2.f * a);
+    float t2 = (-b + sqrtf(discriminant)) / (2.f * a);
 
     return Intersections(
         Intersection(t1, shared_from_this()),
